Add `tb clone` to TB Clone Repos From Terminal
Add the command `tb clone` to clone repos right from the terminal
instead of having to go to GitHub, find the repo that the developer
wants to clone to go the terminal and then typing
`git clone {Repo URL}`.

This command would speed up the development process by skipping over the
going to the browser to get the repos URL and instead have the developer
type in the service that they want to start developing from the
terminal.